Move block cache related code into separate file
